fix: change id to uint64

+++ client/subscribe.go
@@ -6,24 +6,24 @@ import (
 )
 
 func (c *Client) validateSubscribing(req Request) error {
-  if req.Method != methodSubscribe {
-    return fmt.Errorf("invalid method")
-  }
+	if req.Method != methodSubscribe {
+		return fmt.Errorf("invalid method")
+	}
 
-  if len(req.Params) == 0 {
-    return fmt.Errorf("params are required")
-  }
+	if len(req.Params) == 0 {
+		return fmt.Errorf("params are required")
+	}
 
-  for _, event := range req.Params {
-    if _, ok := validSubscriptionEvents[event]; !ok {
-      return fmt.Errorf("invalid params")
-    }
-  }
+	for _, event := range req.Params {
+		if _, ok := validSubscriptionEvents[event]; !ok {
+			return fmt.Errorf("invalid params")
+		}
+	}
 
-  return nil
+	return nil
 }
 
-func newSubscribingResponse(id int) ([]byte, error) {
+func newSubscribingResponse(id uint64) ([]byte, error) {
 	bb, err := json.Marshal(Response{
 		ID:     id,
 		Result: nil,
@@ -34,4 +34,3 @@ func newSubscribingResponse(id int) ([]byte, error) {
 
 	return bb, nil
 }
-
